+++ ricaun.Revit.UI/AppLoaderAttribute.cs
@@ -5,6 +5,7 @@ namespace ricaun.Revit.UI
     /// <summary>
     /// Console Attribute for ricaun.AppLoader
     /// </summary>
+    [Obsolete("This method gonna be removed, use AppLoaderAttribute")]
     public class ConsoleAttribute : Attribute
     {
     }