+++ ricaun.Revit.UI/RibbonSplitExtension.cs
@@ -1,4 +1,5 @@
 ﻿using Autodesk.Revit.UI;
+using System;
 using System.Linq;
 
 namespace ricaun.Revit.UI
@@ -13,11 +14,11 @@ namespace ricaun.Revit.UI
         /// CreateSplitButton
         /// </summary>
         /// <param name="ribbonPanel"></param>
-        /// <param name="targetPushButtons"></param>
+        /// <param name="pushButtons"></param>
         /// <returns></returns>
-        public static SplitButton CreateSplitButton(this RibbonPanel ribbonPanel, params PushButtonData[] targetPushButtons)
+        public static SplitButton CreateSplitButton(this RibbonPanel ribbonPanel, params PushButtonData[] pushButtons)
         {
-            return ribbonPanel.CreateSplitButton(null, targetPushButtons);
+            return ribbonPanel.CreateSplitButton(null, pushButtons);
         }
 
         /// <summary>
@@ -25,32 +26,92 @@ namespace ricaun.Revit.UI
         /// </summary>
         /// <param name="ribbonPanel"></param>
         /// <param name="targetText"></param>
-        /// <param name="targetPushButtons"></param>
+        /// <param name="pushButtons"></param>
         /// <returns></returns>
-        public static SplitButton CreateSplitButton(this RibbonPanel ribbonPanel, string targetText, params PushButtonData[] targetPushButtons)
+        public static SplitButton CreateSplitButton(this RibbonPanel ribbonPanel, string targetText, params PushButtonData[] pushButtons)
         {
-            SplitButton currentSplitButton = null;
-            if (targetPushButtons.Any())
-            {
-                if (targetText == null) targetText = targetPushButtons.FirstOrDefault().Text;
-                var targetName = targetText;
+            SplitButton splitButton = null;
 
-                while (RibbonSafeExtension.VerifyNameExclusive(ribbonPanel, targetName))
-                    targetName = RibbonSafeExtension.SafeButtonName(targetText);
+            if (targetText is null)
+                targetText = pushButtons.FirstOrDefault()?.Text ?? nameof(SplitButton);
 
+            var targetName = targetText;
 
-                currentSplitButton = ribbonPanel.AddItem(new SplitButtonData(targetName, targetText)) as SplitButton;
+            targetName = RibbonSafeExtension.GenerateSafeButtonName(ribbonPanel, targetName, targetText);
 
-                foreach (PushButtonData currentPushButton in targetPushButtons)
-                {
-                    while (RibbonSafeExtension.VerifyNameExclusive(currentSplitButton, currentPushButton.Name))
-                        currentPushButton.Name = RibbonSafeExtension.SafeButtonName(targetText);
+            splitButton = ribbonPanel.AddItem(new SplitButtonData(targetName, targetText)) as SplitButton;
 
-                    currentSplitButton.AddPushButton(currentPushButton);
-                }
-            }
+            splitButton.AddPushButtons(pushButtons);
 
-            return currentSplitButton;
+            return splitButton;
+        }
+        #endregion
+
+        #region CreatePushButton
+        /// <summary>
+        /// CreatePushButton
+        /// </summary>
+        /// <typeparam name="TExternalCommand"></typeparam>
+        /// <param name="splitButton"></param>
+        /// <param name="text"></param>
+        /// <returns></returns>
+        public static PushButton CreatePushButton<TExternalCommand>(this SplitButton splitButton, string text = null) where TExternalCommand : class, IExternalCommand, new()
+        {
+            PushButton pushButton = splitButton.AddPushButton(splitButton.NewPushButtonData<TExternalCommand>(text)) as PushButton;
+            return pushButton;
+        }
+        /// <summary>
+        /// CreatePushButton
+        /// </summary>
+        /// <typeparam name="TExternalCommand"></typeparam>
+        /// <typeparam name="TAvailability"></typeparam>
+        /// <param name="splitButton"></param>
+        /// <param name="text"></param>
+        /// <returns></returns>
+        public static PushButton CreatePushButton<TExternalCommand, TAvailability>(this SplitButton splitButton, string text = null) where TExternalCommand : class, IExternalCommand, new() where TAvailability : class, IExternalCommandAvailability, new()
+        {
+            PushButton pushButton = splitButton
+                .CreatePushButton<TExternalCommand>(text)
+                .SetAvailability<TAvailability>();
+
+            return pushButton;
+        }
+        #endregion
+
+        #region NewPushButtonData
+        /// <summary>
+        /// NewPushButtonData
+        /// </summary>
+        /// <param name="splitButton"></param>
+        /// <param name="commandType"></param>
+        /// <param name="text"></param>
+        /// <returns></returns>
+        public static PushButtonData NewPushButtonData(this SplitButton splitButton, Type commandType, string text = null)
+        {
+            return RibbonSafeExtension.NewPushButtonData(splitButton, commandType, text);
+        }
+        /// <summary>
+        /// NewPushButtonData
+        /// </summary>
+        /// <typeparam name="TExternalCommand"></typeparam>
+        /// <param name="splitButton"></param>
+        /// <param name="text"></param>
+        /// <returns></returns>
+        public static PushButtonData NewPushButtonData<TExternalCommand>(this SplitButton splitButton, string text = null) where TExternalCommand : class, IExternalCommand, new()
+        {
+            return RibbonSafeExtension.NewPushButtonData<TExternalCommand>(splitButton, text);
+        }
+        /// <summary>
+        /// NewPushButtonData
+        /// </summary>
+        /// <typeparam name="TExternalCommand"></typeparam>
+        /// <typeparam name="TAvailability"></typeparam>
+        /// <param name="splitButton"></param>
+        /// <param name="text"></param>
+        /// <returns></returns>
+        public static PushButtonData NewPushButtonData<TExternalCommand, TAvailability>(this SplitButton splitButton, string text = null) where TExternalCommand : class, IExternalCommand, new() where TAvailability : class, IExternalCommandAvailability, new()
+        {
+            return RibbonSafeExtension.NewPushButtonData<TExternalCommand, TAvailability>(splitButton, text);
         }
         #endregion
     }