+++ ricaun.Revit.UI/RibbonTabExtension.cs
@@ -11,7 +11,6 @@ namespace ricaun.Revit.UI
     public static class RibbonTabExtension
     {
         #region Select
-
         /// <summary>
         /// GetRibbonTab
         /// </summary>
@@ -25,19 +24,12 @@ namespace ricaun.Revit.UI
         /// <summary>
         /// GetRibbonTab
         /// </summary>
-        /// <param name="tabId"></param>
+        /// <param name="ribbonTabId"></param>
         /// <returns></returns>
-        public static Autodesk.Windows.RibbonTab GetRibbonTab(string tabId)
+        public static Autodesk.Windows.RibbonTab GetRibbonTab(string ribbonTabId)
         {
             var ribbon = Autodesk.Windows.ComponentManager.Ribbon;
-            foreach (Autodesk.Windows.RibbonTab tab in ribbon.Tabs)
-            {
-                if (tab.Id == tabId)
-                {
-                    return tab;
-                }
-            }
-            return null;
+            return ribbon.FindTab(ribbonTabId);
         }
 
         /// <summary>
@@ -60,9 +52,23 @@ namespace ricaun.Revit.UI
         public static bool Remove(this Autodesk.Windows.RibbonTab ribbonTab)
         {
             var ribbon = Autodesk.Windows.ComponentManager.Ribbon;
-            GetRibbonTabsDictionary()?.Remove(ribbonTab.Name);
+            GetRibbonTabsDictionary()?.Remove(ribbonTab.Id);
             return ribbon.Tabs.Remove(ribbonTab);
         }
+
+        /// <summary>
+        /// Remove RibbonPanel / Remove Revit Dictionary Name
+        /// </summary>
+        /// <param name="ribbonTab"></param>
+        /// <param name="ribbonPanel"></param>
+        /// <returns></returns>
+        public static bool Remove(this Autodesk.Windows.RibbonTab ribbonTab, Autodesk.Windows.RibbonPanel ribbonPanel)
+        {
+            var removed = ribbonTab.Panels.Remove(ribbonPanel);
+            GetRibbonTabsDictionary(ribbonTab)?.Remove(ribbonPanel.Source.Name);
+            return removed;
+        }
+
         #endregion
 
         #region Order
@@ -113,8 +119,29 @@ namespace ricaun.Revit.UI
         /// <summary>
         /// GetRibbonTabsDictionary
         /// </summary>
+        /// <param name="ribbonTab"></param>
         /// <returns></returns>
-        private static Dictionary<string, Dictionary<string, RibbonPanel>> GetRibbonTabsDictionary()
+        internal static Dictionary<string, RibbonPanel> GetRibbonTabsDictionary(Autodesk.Windows.RibbonTab ribbonTab)
+        {
+            return GetRibbonTabsDictionary(ribbonTab.Id);
+        }
+        /// <summary>
+        /// GetRibbonTabsDictionary
+        /// </summary>
+        /// <param name="ribbonTabId"></param>
+        /// <returns></returns>
+        internal static Dictionary<string, RibbonPanel> GetRibbonTabsDictionary(string ribbonTabId)
+        {
+            if (GetRibbonTabsDictionary().TryGetValue(ribbonTabId, out Dictionary<string, RibbonPanel> value))
+                return value;
+
+            return null;
+        }
+        /// <summary>
+        /// GetRibbonTabsDictionary
+        /// </summary>
+        /// <returns></returns>
+        internal static Dictionary<string, Dictionary<string, RibbonPanel>> GetRibbonTabsDictionary()
         {
             var type = typeof(UIApplication);